Create README.md
This project consist of a simple Power BI file that is connected to a SQL Server Database (version 17). There is a database file (OEE.bak) that can be restore to your local SQL Server Instance (use SSMS). In addition there is an PDF file that contains that database diagram extracted from SSMS.